backend/requirements.txt
alembic==1.16.5
eventlet==0.40.3
fastjsonschema==2.21.1
alembic==1.16.5
Flask==3.1.2
flask-cors==6.0.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server-mathjax==0.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_git==0.51.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
lizard==1.18.0
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.3
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
narwhals==1.46.0
nbclient==0.10.2
nbconvert==7.16.6
nbdime==4.0.2
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook_shim==0.2.4
psycopg2-binary==2.9.11
PyJWT==2.10.1
pyparsing==3.2.3
python-can==4.6.1
python-dateutil==2.9.0.post0
python-engineio==4.12.3
python-engineio==4.12.3
python-socketio==5.14.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
requests==2.32.4
SQLAlchemy==2.0.43
Werkzeug==3.1.3
Werkzeug==3.1.3
paho-mqtt==1.6.1